+++ src/index.ts
@@ -1,8 +1,8 @@
-export * from './leaflet-adapter/circle.js'
-export * from './leaflet-adapter/coordinate-reference-system/epsg-3857.js'
-export * from './leaflet-adapter/map/map.js'
-export * from './leaflet-adapter/marker.js'
-export * from './leaflet-adapter/polygon.js'
-export * from './leaflet-adapter/popup.js'
-export * from './leaflet-adapter/tile-layer/tile-layer-osm.js'
-export * from './leaflet-adapter/tile-layer/tile-layer.js'
+export * from './leaf/circle.js'
+export * from './leaf/coordinate-reference-system/epsg-3857.js'
+export * from './leaf/map/map.js'
+export * from './leaf/marker.js'
+export * from './leaf/polygon.js'
+export * from './leaf/popup.js'
+export * from './leaf/tile-layer/tile-layer-osm.js'
+export * from './leaf/tile-layer/tile-layer.js'
